src/main/java/recipeproject/cookbook/controllers/IndexController.java
package recipeproject.cookbook.controllers;

import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.RequestMapping;
import recipeproject.cookbook.service.RecipeService;
import recipeproject.cookbook.service.RecipeServiceImpl;



@Controller
public class IndexController {

    private final RecipeService recipeService;

    public IndexController(RecipeService recipeService) {
        this.recipeService = recipeService;
    }

    @RequestMapping({"", "/", "/index"})
    public String getIndexPage(Model model){

        model.addAttribute("recipes", recipeService.getRecipies());

        return "index";
    }
}
